feat : 회원가입할 때 role 추가 - Integer로 자료형 변경

+++ src/main/java/koreatechbus/dto/user/RegisterDTO.java
@@ -5,7 +5,7 @@ public record RegisterDTO(
     String name,
     String password,
     String email,
-    Long role
+    Integer role
 ) {
 
 }
